TST: create LiveDispatcher test with reactivex over streamz.
- refactor tests that are solely exemplary to integration_examples
